Rename Untitled.ipynb to Part1.ipynb
Canadian Tire wants to start a word-of-mouth marketing campaign over Instagram for their new products. To do so, they need to select users and pay them to advertise the products. After careful investigation of Instagram users, they could come up with a network of potential users and their connections. However, due to the budget limitation they cannot ask all users to advertise the product. Therefore, they are looking for a solution to be able to only select 4 users to advertise the products. The following network (graph) represents users and the connections between users. In this network, each node is a user and each arrow show if a user i follows user j on Instagram. For example, user 𝑈1 follows user 𝑈3.

Given the graph data, implement a python program for the following steps.
1) [2 point] Import this graph to Python and visualize the graph.
2) [3 point] Using the PageRank Algorithm find top-4 important users that Canadian Tire can target for advertising. Explain why you think these selected users are the best ones to be targeted.